+++ Package.swift
@@ -13,9 +13,10 @@ let package = Package(
     dependencies: [
         .package(url: "https://github.com/JohnSundell/Marathon.git", from: "3.1.0"),
         .package(url: "https://github.com/JohnSundell/ShellOut.git", from: "2.1.0"),
+        .package(url: "https://github.com/nerdishbynature/octokit.swift", from: "0.9.0")
     ],
     targets: [
-        .target(name: "Danger", dependencies: ["ShellOut"]),
+        .target(name: "Danger", dependencies: ["ShellOut", "OctoKit"]),
         .target(name: "Runner", dependencies: ["Danger", "MarathonCore"]),
         .testTarget(name: "DangerTests", dependencies: ["Danger"]),
     ],